+++ src/main.js
@@ -1,13 +1,13 @@
-import { fetchData } from "./lib/api.js";
+import { createCharacter } from "./lib/character.js";
 
 const fetchButton = document.querySelector(".header__search-button");
+
+fetchButton.addEventListener("click", () => {
+  clearCharacters();
+  createCharacter();
+});
 
 function clearCharacters() {
   const characterElements = document.querySelectorAll(".character");
   characterElements.forEach((section) => section.remove());
 }
-
-fetchButton.addEventListener("click", () => {
-  clearCharacters();
-  fetchData();
-});
